pyproject.toml
[project]
name = "adk-agent-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adktools>=0.1.0",
    "google-adk>=0.1.0",
    "litellm>=1.65.8",
    "pip>=25.0.1",
]
